Better functionality for FunctionCall
